+++ relay-server/src/utils/managed_envelope.rs
@@ -421,6 +421,7 @@ impl ManagedEnvelope {
                     tags.has_sessions = summary.session_quantity > 0,
                     tags.has_profiles = summary.profile_quantity > 0,
                     tags.has_transactions = summary.secondary_transaction_quantity > 0,
+                    tags.has_span_metrics = summary.secondary_span_quantity > 0,
                     tags.has_replays = summary.replay_quantity > 0,
                     tags.has_checkins = summary.checkin_quantity > 0,
                     tags.event_category = ?summary.event_category,
@@ -459,6 +460,17 @@ impl ManagedEnvelope {
             );
         }
 
+        if self.context.summary.span_quantity > 0 {
+            self.track_outcome(
+                outcome.clone(),
+                match self.context.summary.span_metrics_extracted {
+                    true => DataCategory::SpanIndexed,
+                    false => DataCategory::Span,
+                },
+                self.context.summary.span_quantity,
+            );
+        }
+
         // Track outcomes for attached secondary transactions, e.g. extracted from metrics.
         //
         // Primary transaction count is already tracked through the event category
@@ -469,6 +481,18 @@ impl ManagedEnvelope {
                 // Secondary transaction counts are never indexed transactions
                 DataCategory::Transaction,
                 self.context.summary.secondary_transaction_quantity,
+            );
+        }
+
+        // Track outcomes for attached secondary spans, e.g. extracted from metrics.
+        //
+        // Primary span count is already tracked through `SpanIndexed`.
+        if self.context.summary.secondary_span_quantity > 0 {
+            self.track_outcome(
+                outcome.clone(),
+                // Secondary transaction counts are never indexed transactions
+                DataCategory::Span,
+                self.context.summary.secondary_span_quantity,
             );
         }
 
@@ -551,5 +575,47 @@ impl Drop for ManagedEnvelope {
 impl<G> From<TypedEnvelope<G>> for ManagedEnvelope {
     fn from(value: TypedEnvelope<G>) -> Self {
         value.0
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use bytes::Bytes;
+
+    #[test]
+    fn span_metrics_are_reported() {
+        let bytes =
+            Bytes::from(r#"{"dsn":"https://e12d836b15bb49d7bbf99e64295d995b:@sentry.io/42"}"#);
+        let envelope = Envelope::parse_bytes(bytes).unwrap();
+
+        let (test_store, _) = Addr::custom();
+        let (outcome_aggregator, mut rx) = Addr::custom();
+        let mut env = ManagedEnvelope::new_internal(
+            envelope,
+            None,
+            outcome_aggregator,
+            test_store,
+            ProcessingGroup::Ungrouped,
+        );
+        env.context.summary.span_metrics_extracted = true;
+        env.context.summary.span_quantity = 123;
+        env.context.summary.secondary_span_quantity = 456;
+
+        env.reject(Outcome::Abuse);
+
+        rx.close();
+
+        let outcome = rx.blocking_recv().unwrap();
+        assert_eq!(outcome.category, DataCategory::SpanIndexed);
+        assert_eq!(outcome.quantity, 123);
+        assert_eq!(outcome.outcome, Outcome::Abuse);
+
+        let outcome = rx.blocking_recv().unwrap();
+        assert_eq!(outcome.category, DataCategory::Span);
+        assert_eq!(outcome.quantity, 456);
+        assert_eq!(outcome.outcome, Outcome::Abuse);
+
+        assert!(rx.blocking_recv().is_none());
     }
 }